Move 'core' signals to signals.py
